Added new example terraform that deploys an ASG with one node for our log snapshot test Updated ec2-files.go to incorporate PR comments Separated the scp test cases from the ssh test. Scp test cases now live in their own test.
