+++ examples/terraform-asg-scp-example/main.tf
@@ -1,0 +1,96 @@
+# ---------------------------------------------------------------------------------------------------------------------
+# DEPLOY TWO EC2 INSTANCES THAT ALLOWS CONNECTIONS VIA SSH
+# See test/terraform_ssh_example.go for how to write automated tests for this code.
+# ---------------------------------------------------------------------------------------------------------------------
+
+provider "aws" {
+  region = "${var.aws_region}"
+}
+
+# ---------------------------------------------------------------------------------------------------------------------
+# DEPLOY AN ASG WITH ONE NODE TO TEST HOW WE CAN SCP FROM THE EC2 INSTANCE IN THIS ASG
+# ---------------------------------------------------------------------------------------------------------------------
+
+resource "aws_launch_template" "sample_launch_template" {
+  name_prefix            = "${var.instance_name}"
+  image_id               = "${data.aws_ami.ubuntu.id}"
+  instance_type          = "t2.micro"
+  vpc_security_group_ids = ["${aws_security_group.example.id}"]
+  key_name               = "${var.key_pair_name}"
+}
+
+resource "aws_autoscaling_group" "sample_asg" {
+  vpc_zone_identifier = ["${data.aws_subnet_ids.default_subnets.ids}"]
+  availability_zones  = []
+
+  desired_capacity = 1
+  max_size         = 1
+  min_size         = 1
+
+  launch_template = {
+    id      = "${aws_launch_template.sample_launch_template.id}"
+    version = "$$Latest"
+  }
+}
+
+# ---------------------------------------------------------------------------------------------------------------------
+# CREATE A SECURITY GROUP TO CONTROL WHAT REQUESTS CAN GO IN AND OUT OF THE EC2 INSTANCES
+# ---------------------------------------------------------------------------------------------------------------------
+
+resource "aws_security_group" "example" {
+  name = "${var.instance_name}"
+
+  egress {
+    from_port   = 0
+    to_port     = 0
+    protocol    = "-1"
+    cidr_blocks = ["0.0.0.0/0"]
+  }
+
+  ingress {
+    from_port = "${var.ssh_port}"
+    to_port   = "${var.ssh_port}"
+    protocol  = "tcp"
+
+    # To keep this example simple, we allow incoming SSH requests from any IP. In real-world usage, you should only
+    # allow SSH requests from trusted servers, such as a bastion host or VPN server.
+    cidr_blocks = ["0.0.0.0/0"]
+  }
+}
+
+# ---------------------------------------------------------------------------------------------------------------------
+# LOOK UP THE LATEST UBUNTU AMI
+# ---------------------------------------------------------------------------------------------------------------------
+
+data "aws_ami" "ubuntu" {
+  most_recent = true
+  owners      = ["099720109477"] # Canonical
+
+  filter {
+    name   = "virtualization-type"
+    values = ["hvm"]
+  }
+
+  filter {
+    name   = "architecture"
+    values = ["x86_64"]
+  }
+
+  filter {
+    name   = "image-type"
+    values = ["machine"]
+  }
+
+  filter {
+    name   = "name"
+    values = ["ubuntu/images/hvm-ssd/ubuntu-xenial-16.04-amd64-server-*"]
+  }
+}
+
+data "aws_vpc" "default" {
+  default = true
+}
+
+data "aws_subnet_ids" "default_subnets" {
+  vpc_id = "${data.aws_vpc.default.id}"
+}
