+++ examples/terraform-asg-scp-example/outputs.tf
@@ -1,0 +1,3 @@
+output "asg_name" {
+  value = "${aws_autoscaling_group.sample_asg.name}"
+}
